+++ CHANGES.rst
@@ -14,7 +14,8 @@ Changes:
 
 Fixes:
 ------
-- No change.
+- Update ``mimeType`` to ``mediaType`` as format type representation according to `OGC-API`
+  (fixes `#211  <https://github.com/crim-ca/weaver/issues/211>`_).
 
 `3.5.0 <https://github.com/crim-ca/weaver/tree/3.5.0>`_ (2021-08-19)
 ========================================================================